Remove `src/app/gui/app/index.js`
- remove `gui/app/index` module
- move `gui/app/index::ApplicationTemplate` class into `app/main.js`
